Add files via upload
This file explains the order in which programs are run to create the data that will be loaded into the London Stage database.  Also, how to extract the data out of the works and author spreadsheets, and loading that data into the appropriate database files.  The documentation explains specific keys to create in the database and creating and cleaning new fields used for searching in the database.